created userModel, began abstracting the clients model into it
